+++ src/interfaces/UsdnProtocol/IUsdnProtocol.sol
@@ -96,23 +96,23 @@ interface IUsdnProtocol is IUsdnProtocolActions {
      * @notice Set the soft long expo imbalance limit percentage
      * @param newLimit The new percentage limit (based on EXPO_IMBALANCE_LIMIT_DENOMINATOR)
      */
-    function setSoftLongExpoImbalanceLimit(uint16 newLimit) external;
+    function setSoftLongExpoImbalanceLimit(int256 newLimit) external;
 
     /**
      * @notice Set the hard long expo imbalance limit percentage
      * @param newLimit The new percentage limit (based on EXPO_IMBALANCE_LIMIT_DENOMINATOR)
      */
-    function setHardLongExpoImbalanceLimit(uint16 newLimit) external;
+    function setHardLongExpoImbalanceLimit(int256 newLimit) external;
 
     /**
      * @notice Set the soft vault expo imbalance limit percentage
      * @param newLimit The new percentage limit (based on EXPO_IMBALANCE_LIMIT_DENOMINATOR)
      */
-    function setSoftVaultExpoImbalanceLimit(uint16 newLimit) external;
+    function setSoftVaultExpoImbalanceLimit(int256 newLimit) external;
 
     /**
      * @notice Set the hard vault expo imbalance limit percentage
      * @param newLimit The new percentage limit (based on EXPO_IMBALANCE_LIMIT_DENOMINATOR)
      */
-    function setHardVaultExpoImbalanceLimit(uint16 newLimit) external;
+    function setHardVaultExpoImbalanceLimit(int256 newLimit) external;
 }
